+++ src/main/java/com/mert/taskmanager/service/concretes/TaskManager.java
@@ -1,5 +1,7 @@
 package com.mert.taskmanager.service.concretes;
 
+import com.mert.taskmanager.config.SecurityUtils;
+import com.mert.taskmanager.core.exceptions.AccessDeniedException;
 import com.mert.taskmanager.core.exceptions.ResourceNotFoundException;
 import com.mert.taskmanager.dto.request.task.TaskSaveRequest;
 import com.mert.taskmanager.dto.request.task.TaskUpdateRequest;
@@ -7,6 +9,7 @@ import com.mert.taskmanager.dto.response.TaskResponse;
 import com.mert.taskmanager.entity.Project;
 import com.mert.taskmanager.entity.Task;
 import com.mert.taskmanager.core.mapper.TaskMapper;
+import com.mert.taskmanager.entity.User;
 import com.mert.taskmanager.repository.ProjectRepo;
 import com.mert.taskmanager.repository.TaskRepo;
 import com.mert.taskmanager.service.abstracts.ITaskService;
@@ -38,31 +41,49 @@ public class TaskManager implements ITaskService {
 
     @Override
     public TaskResponse save(TaskSaveRequest taskSaveRequest) {
+        User currentUser= SecurityUtils.getCurrentUser();
+
         Task saveTask = taskMapper.toEntity(taskSaveRequest);
         Project project= projectRepo.findById(taskSaveRequest.getProjectId()).orElseThrow(()->new ResourceNotFoundException(Msg.TASK_PROJECT_NOTFOUND));
-        saveTask.setProject(project);
-        taskRepo.save(saveTask);
-        TaskResponse taskResponse =taskMapper.toResponse(saveTask);
-        return taskResponse;
-
+        if (currentUser.getId().equals(project.getUser().getId())) {
+            saveTask.setProject(project);
+            taskRepo.save(saveTask);
+            TaskResponse taskResponse =taskMapper.toResponse(saveTask);
+            return taskResponse;
+        }
+        else throw new AccessDeniedException("Bu projeye erişme yetkiniz bulunmamaktadır.");
     }
 
     @Override
     public TaskResponse get(Long id) {
+        User currentUser= SecurityUtils.getCurrentUser();
+        Long currentUserId=currentUser.getId();
+
         Task getTask=taskRepo.findById(id).orElseThrow(()->new ResourceNotFoundException(Msg.TASK_NOTFOUND));
-        TaskResponse taskResponse=taskMapper.toResponse(getTask);
-        return  taskResponse;
+        if (currentUserId.equals(getTask.getProject().getUser().getId())) {
+            TaskResponse taskResponse=taskMapper.toResponse(getTask);
+            return  taskResponse;
+        }
+        else throw new AccessDeniedException("Bu projeye erişme yetkiniz bulunmamaktadır.");
     }
 
     @Override
     public TaskResponse update(TaskUpdateRequest taskUpdateRequest) {
+        User currentUser= SecurityUtils.getCurrentUser();
+        Long currentUserId=currentUser.getId();
+
         Task updatedTask=taskMapper.toEntity(taskUpdateRequest);
         Task oldTask=taskRepo.findById(taskUpdateRequest.getId()).orElseThrow(()->new ResourceNotFoundException(Msg.TASK_NOTFOUND));
         Project project=projectRepo.findById(taskUpdateRequest.getProjectId()).orElseThrow(()->new ResourceNotFoundException(Msg.TASK_PROJECT_NOTFOUND));
-        updatedTask.setProject(project);
-        taskRepo.save(updatedTask);
-        TaskResponse taskResponse =taskMapper.toResponse(updatedTask);
-        return  taskResponse;
+
+        if (currentUserId.equals(project.getUser().getId())) {
+            updatedTask.setProject(project);
+            taskRepo.save(updatedTask);
+            TaskResponse taskResponse =taskMapper.toResponse(updatedTask);
+            return  taskResponse;
+
+        }
+        else throw new AccessDeniedException("Bu projeye erişme yetkiniz bulunmamaktadır.");
     }
 
     @Override//Belki kullanırım
@@ -71,19 +92,32 @@ public class TaskManager implements ITaskService {
         if (projectId == null) {
             throw new IllegalArgumentException(Msg.VALIDATE_PROJECT_ID_NULL);
         }
+
+        User currentUser= SecurityUtils.getCurrentUser();
+        Long currentUserId=currentUser.getId();
+        Project project=projectRepo.findById(projectId).orElseThrow(()->new ResourceNotFoundException(Msg.PROJECT_NOTFOUND));//Login olan id ile projectten user id yi almak için çağrdım
+
         Pageable pageable = PageRequest.of(page,pageSize);
         Page<Task> taskPage = taskRepo.findByProjectId(projectId,pageable);
         List<TaskResponse> taskResponseList=new ArrayList<>();
-        for (Task task :taskPage.getContent()){
+        if (currentUserId.equals(project.getUser().getId())) {
+            for (Task task :taskPage.getContent()){
                 taskResponseList.add(taskMapper.toResponse(task));
+            }
+            return new PageImpl<>(taskResponseList,pageable,taskPage.getTotalElements());
         }
-        return new PageImpl<>(taskResponseList,pageable,taskPage.getTotalElements());
+        else throw new AccessDeniedException("Bu projeye erişme yetkiniz bulunmamaktadır.");
     }
 
     @Override
     public boolean delete(Long id) {
+        User currentUser= SecurityUtils.getCurrentUser();
+        Long currentUserId=currentUser.getId();
         Task deleteTask=taskRepo.findById(id).orElseThrow(()->new ResourceNotFoundException(Msg.TASK_NOTFOUND));
-        taskRepo.delete(deleteTask);
-        return  true;
+        if (currentUserId.equals(deleteTask.getProject().getUser().getId())) {
+            taskRepo.delete(deleteTask);
+            return  true;
+        }
+        else throw new AccessDeniedException("Bu projeye erişme yetkiniz bulunmamaktadır.");
     }
 }
